# quadracode-agent-registry/openapi.json
{"openapi":"3.1.0","info":{"title":"Quadracode Agent Registry","description":"Lightweight registry for Quadracode agents (SQLite-backed)","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/agents/register":{"post":{"summary":"Register Agent","operationId":"register_agent_agents_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentRegistrationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agents/{agent_id}/heartbeat":{"post":{"summary":"Agent Heartbeat","operationId":"agent_heartbeat_agents__agent_id__heartbeat_post","parameters":[{"name":"agent_id","in":"path","required":true,"schema":{"type":"string","title":"Agent Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentHeartbeat"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agents":{"get":{"summary":"List Agents","operationId":"list_agents_agents_get","parameters":[{"name":"healthy_only","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Healthy Only"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agents/{agent_id}":{"get":{"summary":"Get Agent","operationId":"get_agent_agents__agent_id__get","parameters":[{"name":"agent_id","in":"path","required":true,"schema":{"type":"string","title":"Agent Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Unregister Agent","operationId":"unregister_agent_agents__agent_id__delete","parameters":[{"name":"agent_id","in":"path","required":true,"schema":{"type":"string","title":"Agent Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stats":{"get":{"summary":"Get Stats","operationId":"get_stats_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistryStats"}}}}}}}},"components":{"schemas":{"AgentHeartbeat":{"properties":{"agent_id":{"type":"string","title":"Agent Id","description":"Agent identifier sending the heartbeat"},"status":{"$ref":"#/components/schemas/AgentStatus","description":"Reported health status","default":"healthy"},"reported_at":{"type":"string","format":"date-time","title":"Reported At","description":"Heartbeat timestamp"}},"type":"object","required":["agent_id"],"title":"AgentHeartbeat","description":"Heartbeat payload reported by agents to indicate liveness."},"AgentInfo":{"properties":{"agent_id":{"type":"string","title":"Agent Id"},"host":{"type":"string","title":"Host"},"port":{"type":"integer","title":"Port"},"status":{"$ref":"#/components/schemas/AgentStatus"},"registered_at":{"type":"string","format":"date-time","title":"Registered At"},"last_heartbeat":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Heartbeat"}},"type":"object","required":["agent_id","host","port","status","registered_at","last_heartbeat"],"title":"AgentInfo","description":"Full record for an agent maintained by the registry service."},"AgentListResponse":{"properties":{"agents":{"items":{"$ref":"#/components/schemas/AgentInfo"},"type":"array","title":"Agents"},"healthy_only":{"type":"boolean","title":"Healthy Only","default":false}},"type":"object","required":["agents"],"title":"AgentListResponse","description":"Response envelope returning a set of agents to callers."},"AgentRegistrationRequest":{"properties":{"agent_id":{"type":"string","title":"Agent Id","description":"Unique identifier for the agent"},"host":{"type":"string","title":"Host","description":"Hostname or IP address reachable by the orchestrator"},"port":{"type":"integer","title":"Port","description":"Primary service port exposed by the agent"}},"type":"object","required":["agent_id","host","port"],"title":"AgentRegistrationRequest","description":"Payload sent by agents when registering with the registry service."},"AgentStatus":{"type":"string","enum":["healthy","unhealthy"],"title":"AgentStatus"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RegistryStats":{"properties":{"total_agents":{"type":"integer","title":"Total Agents"},"healthy_agents":{"type":"integer","title":"Healthy Agents"},"unhealthy_agents":{"type":"integer","title":"Unhealthy Agents"},"last_updated":{"type":"string","format":"date-time","title":"Last Updated"}},"type":"object","required":["total_agents","healthy_agents","unhealthy_agents"],"title":"RegistryStats","description":"Aggregate statistics exposed by the registry service."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}